doubled the result of the magic percentage and average level calculation as it might end in a decimal.

+++ C#Final/C#Final/Program.cs
@@ -138,8 +138,8 @@
 
 							
 							
-							double averageLevel = totalLevel / characterList.Count;
-							double magicPercentage = (magicCount / characterList.Count) * 100;
+							double averageLevel = (double)totalLevel / characterList.Count;
+							double magicPercentage = ((double)magicCount / characterList.Count) * 100;
 
 							Console.WriteLine($"Total Characters: {characterList.Count}");
 							Console.WriteLine($"Average Level: {averageLevel}");
